Now 0.8.x branch reached state of ALPHA!
GMail Counter 0.8 is now working, but it completely leak of the Head Viewer, if you click on toolbar button nothing happens and many settings are missing.

However GMail.js is now complete (I think) and the load time (the time that pass from when Safari load the extension to when the extension is completely loaded and running) is now lower than 1 second!

0.8 is a completely rewrited and rebuilt from scratch version of GMail Counter and fix some conceptual bug and code ugliness of previous versions.