Cleaned UTF-8 files from BOMs.
BOMs are evil!

" Some software think it's a good idea to embed a BOM (byte order mark) in the beginning of an UTF-8 file, but it is not, because the file can already be recognized, and placing a BOM in the beginning of a file means placing three bytes in the beginning of the file which a program that use the file may not expect (e.g. the shell interpreter looks for #! as the first two bytes of an executable)."